# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ArtistSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "picture": {
            "default": "",
            "title": "Picture",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "ArtistSummary",
        "type": "object"
      },
      "AuthError": {
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
      "Body_set_avatar_api_me_avatar_put": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_set_avatar_api_me_avatar_put",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "title": "MeUpdate",
        "type": "object"
      },
...
      "PlaybackSettings": {
        "properties": {
          "crossfade_duration_sec": {
            "title": "Crossfade Duration Sec",
            "type": "integer"
          },
          "crossfade_enabled": {
            "title": "Crossfade Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "crossfade_enabled",
          "crossfade_duration_sec"
        ],
        "title": "PlaybackSettings",
        "type": "object"
      },
      "PlaybackSettingsUpdate": {
        "properties": {
          "crossfade_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crossfade Duration Sec"
          },
          "crossfade_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crossfade Enabled"
          }
        },
        "title": "PlaybackSettingsUpdate",
        "type": "object"
      },
...
      "PlaylistSummary": {
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_public": {
            "default": false,
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "track_count": {
            "title": "Track Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "track_count"
        ],
        "title": "PlaylistSummary",
        "type": "object"
      },
...
      "PublicProfile": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistSummary"
            },
            "title": "Playlists",
            "type": "array"
          },
          "top_artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistSummary"
            },
            "title": "Top Artists",
            "type": "array"
          }
        },
        "required": [
          "id",
          "playlists",
          "top_artists"
        ],
        "title": "PublicProfile",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_admin": {
            "default": false,
            "title": "Is Admin",
            "type": "boolean"
          },
          "is_approved": {
            "default": false,
            "title": "Is Approved",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sf_session": {
        "description": "HttpOnly JWT session cookie set by the login or registration endpoint.",
        "in": "cookie",
        "name": "sf_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "LoggeRythm API",
    "version": "1.0.0",
    "x-loggerythm-compatible-contract-versions": [
      "v1"
    ],
    "x-loggerythm-contract-version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me": {
      "delete": {
        "description": "Delete the current user's account and all of their data.\n\nPlaylists (+tracks), likes and follows cascade via the ORM; play events,\nparty memberships, hosted parties and invite codes have no ORM cascade and\nare cleaned up explicitly so no orphaned rows survive.",
        "operationId": "delete_me_api_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "update_me_api_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/me/avatar": {
      "put": {
        "operationId": "set_avatar_api_me_avatar_put",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_avatar_api_me_avatar_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Set Avatar",
        "tags": [
          "profile"
        ]
      }
    },
...
    "/api/me/settings": {
      "get": {
        "operationId": "get_settings_api_me_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "update_settings_api_me_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Update Settings",
        "tags": [
          "profile"
        ]
      }
    },
...
    "/api/users/{user_id}": {
      "get": {
        "operationId": "public_profile_api_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/users/{user_id}/avatar": {
      "get": {
        "operationId": "get_avatar_api_users__user_id__avatar_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Avatar",
        "tags": [
          "profile"
        ]
      }
    },
...
  }
}